Remove legacy TicketService/TicketFormService monoliths
Final cut-over of the backend-first migration. The two legacy service
files are now fully encircled by the repository layer and no longer
referenced by any production code. The only remaining couplings were
non-DB utilities that never belonged in the service files:

- DEFAULT_MAX_OPEN_PER_USER / DEFAULT_REOPEN_WINDOW_HOURS /
  DEFAULT_THREAD_LIMIT  → moved to helpers/constants.py
- get_staff_role_ids()   → extracted to helpers/ticket_staff.py
  (config-only parser, no DB access — keeps the bot a thin client)
- RouteExecutionContext  → extracted to helpers/ticket_route_context.py
  (dataclass + add_answers/is_expired/to_db_dict/from_db_row helpers)

Bot view helpers (ticket_views*, ticket_form_views) repointed to these
neutral modules. Deleted services/ticket_service.py,
services/ticket_form_service.py and their now-redundant unit tests
(logic is covered by backend/db/repository/test_repositories.py).
Removed both files from the modularity ceilings; bumped the tickets
repository ceiling to 1031 for the JSON role-id parser and pagination.

1343 tests pass; ruff, mypy (backend/connectors), and modularity all green.
