+++ tests/test_ticket_form_views.py
@@ -19,7 +19,7 @@ from helpers.ticket_form_views import (
     TicketContinueView,
     create_ticket_from_route,
 )
-from services.ticket_form_service import RouteExecutionContext
+from helpers.ticket_route_context import RouteExecutionContext
 from tests.factories.discord_factories import FakeInteraction
 
 # ---------------------------------------------------------------------------
